+++ ddtrace/contrib/aiopg/__init__.py
@@ -22,7 +22,10 @@ required_modules = ["aiopg"]
 
 with require_modules(required_modules) as missing_modules:
     if not missing_modules:
-        from .patch import get_version
-        from .patch import patch
+        # Required to allow users to import from `ddtrace.contrib.aiohttp.patch` directly
+        from . import patch as _  # noqa: F401, I001
+
+        from ..internal.aiopg.patch import get_version
+        from ..internal.aiopg.patch import patch
 
         __all__ = ["patch", "get_version"]
